move func to inquiry.c
